bugfix: fix password validation on login
When you login, the validator validates the password for the various
conditions for a valid password. The problem with that is a user is
made aware of what exactly is wrong which might be a good thing but
not good when an account is hacked or being brute-forced.

A simple incorrect username/password is enough to inform a user
something's wrong with their credentials.

Completes [HOM-45]