feat: save listing as draft
Be default, a saved listing sets the `available` fields to true. This
means that it is available for leasing and users can search for and
immediately see it. However, there are times where the information
about an apartment is incomplete or wrong and user wants to save it
as draft and continue later to update and upload. User's should be able
to set the `available` field to false so it won't be available for
other users viewing.

Completes [HOM-9]