feat: send verification email
Verification emails are sent to users immediately after registration.
This is to control access and also to prevent spamming so we know
only legitimate people are using the platform.

Whomever is going to make use of the platform must verify the
email used for registration via a link that will be sent to
them.

Completes #[HOM-52]