feat: implement access and refresh token
Token implementation was changed from just access token to incluse
refresh token.

This was done to improve the experience of the user while using
the application. With just access token, when it expires, users
are forced to login again to continue using the app. However,
with the inclusion of refresh token that is not the case as the
refresh token fetches a new access token as it expires and users
won't have to logout and login in again to continue using the app.

Completes [HOM-48]