feat: implement password reset
Password reset was implemented. When users' forget their password
they can send a request to change it by sending the email address
the reset is meant for.

When a request to change password is made, a link is sent to the users'
registered email address.

Also, password change was implemented. Here, user's willingly chose to
change password due to personal reasons.

Completes [HOM-4] [HOM-53]