chore: add spring boot app to docker
The spring boot application was added to docker alongside
postgres and mongodb databases.

Docker allows you to set up your application runtime in such a way
that it runs in exactly the same manner as it will in production,
on the same operating system, with the same environment variables,
and any other configuration and setup you require.

Completes [HOM-62]