chore: make servr async
Make the server asynchronous so some operations don't have
to wait for each other,and in some cases fails just because
the previous operatio failed.

This is obvious during user registration. When a user
registers and there's problem with the mail server an
exception is thrown inspite of the fact that registration
is successful. Failure to send verification email triggers
an exception.

The changes made ensures that mail server problem will not prevent
the user from getting successful registration message. Users can
request to be resent the verification mail.