ref: refactor method name and visibility
Some methods were renamed for better understanding of what they do while
the visibility of some methods were changed.

Methods not used outside of the class were made private. Also, some
methods not present in the interface implemented by a class were added
to the the interface to better cohesion.