ref: refactor change password
Deliberate changing of password was  moved from authentication
service to a newly created user service.

This was done for the purpose of separation of concern. The authentication
service will handle authentication related logic while the user
service will handle user related logic of which deliberate changing
of password is one.