bugfix: resolve time mismatch
Resolve the time mismatch between createdAt view timestamp. There's at least
an hour difference between when  a listing is viewed and when the record is created in
the database. This is due to difference in how time zones are handled in JavaScript and Java.
JavaScript isoString() method return date in UTC while Java LocalDateTime contains no time zone
information.