feat: implement getting views count
As listings get viewed they're recorded in the database. To this end, each listing object
has a viewCount field to it to show how many times it's been viewed.
Also. total views an agent has for his/her listings has been added to the listing stats
on the dashboard. A breakdown of views for each month per year will was also added
to the listing stats.