feat: implement logout on server
Logout was implemented on the server. Token are invalidated and refresh token
deleted. This ensures that even when an attacker get hold of a refresh token
after a user has logged out it cannot be used to get new access token.

Also, an invalidated token cannot be reused; user will have to log in to get
new access token.