feat: manager user session on server
User session is being managed on the server.
Before now, a user can login to the app using as many device as possible.
These changes mitigate that. It forbids a user from logging on using
multiple logins.

Whenever a user logs in, a session token is created. Any further login is
is disallowed so long the user is active on the current session.
Also, a mail is sent to the user to inform them of an existing session
and that a new one wants to be created but disallowed.

Creating a new session while there's an ongoing one could be a sign
of suspicious activity, thus user's are warned in the mail to
swift action to be taken.